reduce surface area with cgo and add invoke test
