+++ libevm/.gitignore
@@ -1,3 +1,3 @@
 bin/
-contracts/compiled/
+interop/compiled/
 test-db
